Refactor Docker configurations and downgrade Dockerfile Maven plugin version
Removes BC-FIPS package configurations, for now. At this time just want this to be very basic and we will add items as needed. Additionally, the Dockerfile Maven plugin version in the pom.xml file is downgraded from 1.4.13 to 1.4.6 because some are having issues with the .13 version on Macs with M1 chips.