Add code that was used for demo to team on 2May2024.  JSON configuration read in to ServiceConfig object by ServiceConfigLoader as application runs. (#3)
When message received via /transform information (organization, inbound endpoint, outbound endpoint) is read from headers to determine the pipelines to run.  Obviously this will change to proper handling.  A ServiceContext is created from this information.  Then a PipelineChain built from the context.  The Pipeline's are then executed on the incoming message.  The transformed message is then submitted through a temporary MllpSender.  The response from the downstream system is then transformed via the Pipeline before sending back to the caller.

A pipeline is made up Request Transformations and Response Transformations.  Each transformation is made up of a precondition operation that must execute true before any changes are made.  Then there is also a chain of Operations which make up the steps executed to transform data.