IGDD-1552 Docker update for keystore (#22)
* Remove hardcoded certificates from the repository.  Generate certificates dynamically during the build for unit testing.

* Update Dockerfile to expose new ports

The Docker configuration has been updated to expose ports 444 and 9082. This change is necessary to align with application yml defaults.

* Update maven.yml to attempt github action unit test fix.

* Add removal of temporary key/trust files from antrun configuration.

* Add transformation service destination to application.yml

The transformation service destination has been added to the application.yml file. It is also now accessed via @Value annotation in the HubController class. This replaces the hardcoded destination URI.