IGDD-2085 - Update bcfips 2.1 (#165)
* Update to use 2.1 versions of BC-FIPS which are now in their own sub-directory.

Dockerfile now copies bc-fips files from their new sub-directory to their new sub-directory on the container.

fatjar-run.sh wildcards the classpath to the bcfips lib directory on the container so that we don't have to manually update in the future when changing bcfips files.

pom.xml removes exclusion for commons-fileupload now that update to izgw-core has been done. Also changed antrun that generates the temporary keystores for testing to locate the bc-fips file dynamically to use the in the keytool commands.

* Fix incorrect forward slash in application.yml for group-role-mapping.

* Update release notes.

* Documentation tweaks to mention that the version of bc-fips can change.

---------

Co-authored-by: Austin Moody <austinmoody@protonmail.com>